requirements.txt
click
enum-compat
enum34
eventlet
Flask
Flask-SocketIO
gevent
gevent-websocket
greenlet
gunicorn
itsdangerous
Jinja2
MarkupSafe
numpy
olefile
opencv-python
Pillow
python-engineio
python-socketio
redis
six
Werkzeug

torch
torchvision
ipython